Add event listener for resize of screen
Added an event listener so that when the user resizes the screen a paragraph appears letting them know how many times they resize it.  HTML was added so that the paragraph text would appear in that spot on the resize, that way not visible to the user if they do not resize the screen.